+++ packages/cloud_firestore/ios/Classes/CloudFirestorePlugin.m
@@ -114,8 +114,8 @@ NSDictionary *parseQuerySnapshot(FIRQuerySnapshot *snapshot) {
       @"type" : type,
       @"document" : documentChange.document.data,
       @"path" : documentChange.document.reference.path,
-      @"oldIndex" : [NSNumber numberWithInt:documentChange.oldIndex],
-      @"newIndex" : [NSNumber numberWithInt:documentChange.newIndex],
+      @"oldIndex" : [NSNumber numberWithUnsignedInteger:documentChange.oldIndex],
+      @"newIndex" : [NSNumber numberWithUnsignedInteger:documentChange.newIndex],
     }];
   }
   return @{
@@ -171,7 +171,7 @@ const UInt8 TIMESTAMP = 136;
   } else if ([value isKindOfClass:[NSData class]]) {
     NSData *blob = value;
     [self writeByte:BLOB];
-    [self writeSize:blob.length];
+    [self writeSize:(UInt32)blob.length];
     [self writeData:blob];
   } else {
     [super writeValue:value];
@@ -295,27 +295,28 @@ const UInt8 TIMESTAMP = 136;
     result(error.flutterError);
   };
   if ([@"Firestore#runTransaction" isEqualToString:call.method]) {
-    [getFirestore(call.arguments) runTransactionWithBlock:^id(FIRTransaction *transaction,
-                                                              NSError **pError) {
-      NSNumber *transactionId = call.arguments[@"transactionId"];
-      NSNumber *transactionTimeout = call.arguments[@"transactionTimeout"];
+    [getFirestore(call.arguments)
+        runTransactionWithBlock:^id(FIRTransaction *transaction, NSError **pError) {
+          NSNumber *transactionId = call.arguments[@"transactionId"];
+          NSNumber *transactionTimeout = call.arguments[@"transactionTimeout"];
 
-      transactions[transactionId] = transaction;
+          self->transactions[transactionId] = transaction;
 
-      dispatch_semaphore_t semaphore = dispatch_semaphore_create(0);
+          dispatch_semaphore_t semaphore = dispatch_semaphore_create(0);
 
-      [self.channel invokeMethod:@"DoTransaction"
-                       arguments:call.arguments
-                          result:^(id doTransactionResult) {
-                            transactionResults[transactionId] = doTransactionResult;
-                            dispatch_semaphore_signal(semaphore);
-                          }];
+          [self.channel invokeMethod:@"DoTransaction"
+                           arguments:call.arguments
+                              result:^(id doTransactionResult) {
+                                self->transactionResults[transactionId] = doTransactionResult;
+                                dispatch_semaphore_signal(semaphore);
+                              }];
 
-      dispatch_semaphore_wait(
-          semaphore, dispatch_time(DISPATCH_TIME_NOW, [transactionTimeout integerValue] * 1000000));
+          dispatch_semaphore_wait(
+              semaphore,
+              dispatch_time(DISPATCH_TIME_NOW, [transactionTimeout integerValue] * 1000000));
 
-      return transactionResults[transactionId];
-    }
+          return self->transactionResults[transactionId];
+        }
         completion:^(id transactionResult, NSError *error) {
           if (error != nil) {
             result([FlutterError errorWithCode:[NSString stringWithFormat:@"%ld", error.code]
@@ -328,7 +329,7 @@ const UInt8 TIMESTAMP = 136;
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
       NSNumber *transactionId = call.arguments[@"transactionId"];
       FIRDocumentReference *document = getDocumentReference(call.arguments);
-      FIRTransaction *transaction = transactions[transactionId];
+      FIRTransaction *transaction = self->transactions[transactionId];
       NSError *error = [[NSError alloc] init];
 
       FIRDocumentSnapshot *snapshot = [transaction getDocument:document error:&error];
@@ -352,7 +353,7 @@ const UInt8 TIMESTAMP = 136;
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
       NSNumber *transactionId = call.arguments[@"transactionId"];
       FIRDocumentReference *document = getDocumentReference(call.arguments);
-      FIRTransaction *transaction = transactions[transactionId];
+      FIRTransaction *transaction = self->transactions[transactionId];
 
       [transaction updateData:call.arguments[@"data"] forDocument:document];
       result(nil);
@@ -361,7 +362,7 @@ const UInt8 TIMESTAMP = 136;
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
       NSNumber *transactionId = call.arguments[@"transactionId"];
       FIRDocumentReference *document = getDocumentReference(call.arguments);
-      FIRTransaction *transaction = transactions[transactionId];
+      FIRTransaction *transaction = self->transactions[transactionId];
 
       [transaction setData:call.arguments[@"data"] forDocument:document];
       result(nil);
@@ -370,7 +371,7 @@ const UInt8 TIMESTAMP = 136;
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
       NSNumber *transactionId = call.arguments[@"transactionId"];
       FIRDocumentReference *document = getDocumentReference(call.arguments);
-      FIRTransaction *transaction = transactions[transactionId];
+      FIRTransaction *transaction = self->transactions[transactionId];
 
       [transaction deleteDocument:document];
       result(nil);