+++ packages/url_launcher/example/lib/main.dart
@@ -34,6 +34,7 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   Future<void> _launched;
+  String _phone = '';
 
   Future<void> _launchInBrowser(String url) async {
     if (await canLaunch(url)) {
@@ -72,6 +73,14 @@ class _MyHomePageState extends State<MyHomePage> {
     }
   }
 
+  Future<void> _makePhoneCall(String url) async {
+    if (await canLaunch(url)) {
+      await launch(url);
+    } else {
+      throw 'Could not launch $url';
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     const String toLaunch = 'https://flutter.io';
@@ -83,6 +92,19 @@ class _MyHomePageState extends State<MyHomePage> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: TextField(
+                  onChanged: (String text) => _phone = text,
+                  decoration: const InputDecoration(
+                      hintText: 'Input the phone number to launch')),
+            ),
+            RaisedButton(
+              onPressed: () => setState(() {
+                    _launched = _makePhoneCall('tel:$_phone');
+                  }),
+              child: const Text('Make phone call'),
+            ),
             const Padding(
               padding: EdgeInsets.all(16.0),
               child: Text(toLaunch),
@@ -106,6 +128,17 @@ class _MyHomePageState extends State<MyHomePage> {
                     _launched = _launchInWebViewWithJavaScript(toLaunch);
                   }),
               child: const Text('Launch in app(JavaScript ON)'),
+            ),
+            const Padding(padding: EdgeInsets.all(16.0)),
+            RaisedButton(
+              onPressed: () => setState(() {
+                    _launched = _launchInWebViewOrVC(toLaunch);
+                    Timer(const Duration(seconds: 5), () {
+                      print('Closing WebView after 5 seconds...');
+                      closeWebView();
+                    });
+                  }),
+              child: const Text('Launch in app + close after 5 seconds'),
             ),
             const Padding(padding: EdgeInsets.all(16.0)),
             FutureBuilder<void>(future: _launched, builder: _launchStatus),