+++ packages/webview_flutter/android/src/main/java/io/flutter/plugins/webviewflutter/FlutterWebView.java
@@ -1,14 +1,13 @@
 package io.flutter.plugins.webviewflutter;
 
-import static io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import static io.flutter.plugin.common.MethodChannel.Result;
-
 import android.content.Context;
 import android.view.View;
 import android.webkit.WebView;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
+import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.platform.PlatformView;
 import java.util.Map;
 
@@ -42,6 +41,24 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
       case "updateSettings":
         updateSettings(methodCall, result);
         break;
+      case "canGoBack":
+        canGoBack(methodCall, result);
+        break;
+      case "canGoForward":
+        canGoForward(methodCall, result);
+        break;
+      case "goBack":
+        goBack(methodCall, result);
+        break;
+      case "goForward":
+        goForward(methodCall, result);
+        break;
+      case "reload":
+        reload(methodCall, result);
+        break;
+      case "currentUrl":
+        currentUrl(methodCall, result);
+        break;
       default:
         result.notImplemented();
     }
@@ -51,6 +68,37 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
     String url = (String) methodCall.arguments;
     webView.loadUrl(url);
     result.success(null);
+  }
+
+  private void canGoBack(MethodCall methodCall, Result result) {
+    result.success(webView.canGoBack());
+  }
+
+  private void canGoForward(MethodCall methodCall, Result result) {
+    result.success(webView.canGoForward());
+  }
+
+  private void goBack(MethodCall methodCall, Result result) {
+    if (webView.canGoBack()) {
+      webView.goBack();
+    }
+    result.success(null);
+  }
+
+  private void goForward(MethodCall methodCall, Result result) {
+    if (webView.canGoForward()) {
+      webView.goForward();
+    }
+    result.success(null);
+  }
+
+  private void reload(MethodCall methodCall, Result result) {
+    webView.reload();
+    result.success(null);
+  }
+
+  private void currentUrl(MethodCall methodCall, Result result) {
+    result.success(webView.getUrl());
   }
 
   @SuppressWarnings("unchecked")
@@ -76,7 +124,7 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
       case 0: // disabled
         webView.getSettings().setJavaScriptEnabled(false);
         break;
-      case 1: //unrestricted
+      case 1: // unrestricted
         webView.getSettings().setJavaScriptEnabled(true);
         break;
       default: