+++ packages/webview_flutter/ios/Classes/WebViewFlutterPlugin.m
@@ -1,0 +1,12 @@
+#import "WebViewFlutterPlugin.h"
+#import "FlutterWebView.h"
+
+@implementation FLTWebViewFlutterPlugin
+
++ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
+  FLTWebViewFactory* webviewFactory =
+      [[FLTWebViewFactory alloc] initWithMessenger:registrar.messenger];
+  [registrar registerViewFactory:webviewFactory withId:@"plugins.flutter.io/webview"];
+}
+
+@end